\name{djia2013}
\docType{data}
\alias{djia2013}
\title{50 Exemplary News Articles from the Reuters-21578 Data Set of Topic acq}
\description{
  The djia2013 dataset contains daily returns for the 30 Dow Jones 
  Industrial Average companies for the year 2013.
}
\usage{data("djia2013")}
\examples{
data("djia2013")
djia2013
}
\keyword{datasets}
